rename postgres_sink to moonlink_sink
