+++ src/moonlink_connectors/src/pg_replicate.rs
@@ -2,7 +2,7 @@
 
 pub mod clients;
 pub mod conversions;
-pub mod postgres_sink;
+pub mod moonlink_sink;
 pub mod postgres_source;
 pub mod replication_state;
 pub mod table;
